// src/services/types/resources/blog.ts
export type BlogapiResponse = {
  _id: string
  __v: number
  blog: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
    }
    section2: {
      image1: string
      image2: string
      image3: string
      image4: string
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      title3_en: string
      title3_fa: string
      title4_en: string
      title4_fa: string
      s_title1_en: string
      s_title1_fa: string
      s_title2_en: string
      s_title2_fa: string
      s_title3_en: string
      s_title3_fa: string
      s_title4_en: string
      s_title4_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      text4_en: string
      text4_fa: string
    }
    section3: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}